+++ src/init-test.js
@@ -2,6 +2,7 @@ import { options } from './config';
 
 
 let jsDomCleanUp = null;
+let sandbox = null;
 
 
 before(function () {
@@ -11,7 +12,17 @@ before(function () {
 });
 
 
+beforeEach(() => {
+    sandbox = sinon.sandbox.create({
+        injectInto: mokamok,
+        properties: ["spy", "stub", "mock", "clock", "server", "requests"],
+        useFakeTimers: true,
+        useFakeServer: true
+    });
+});
+
 afterEach(function () {
+    sandbox.restore();
     if (options.jsdom) {
         jsDomCleanUp();
     }