+++ src/watch.js
@@ -5,6 +5,7 @@ import { removeDependencies, getDependency } from './mokamok';
 import { uncache } from './require-hook';
 import { options } from './config';
 
+
 function updateTest(filePath) {
     removeDependencies(filePath);
     addTestFile(filePath);
@@ -15,8 +16,23 @@ function updateTest(filePath) {
     });
 }
 
+
 export default function watch() {
-    fs.watch('.', { recursive: true }, (eventType, fileName) => {
+    const chokidar = require('chokidar');
+    const dir = path.resolve('.');
+
+    const watchOptions = {
+        persistent: true,
+        ignoreInitial: true,
+    };
+    if (process.platform == 'linux') {
+        watchOptions.awaitWriteFinish = {
+            stabilityThreshold: 2000,
+            pollInterval: 100,
+        };
+    }
+    const watcher = chokidar.watch(dir + '/**/*', watchOptions);
+    watcher.on('all', (event, fileName) => {
         const filePath = path.resolve(fileName);
         if (filePath.indexOf(`/${options.testDirectory}/`) !== -1) {
             uncache(filePath);
@@ -26,7 +42,7 @@ export default function watch() {
             if (testFile) {
                 uncache(filePath);
                 updateTest(testFile);
-             }
+            }
         }
     });
 }