+++ tests/code-coverage/config.js
@@ -4,16 +4,7 @@ module.exports = {
         {
             name: 'with-long-parameter',
             params: '--coverage'
-        }
-    ],
-
-    onStart: function (childProcess) {
-        childProcess.on('done', function () {
-            childProcess._testSnapshot('out', childProcess._stdout);
-        });
-    },
-
-    other: [
+        },
         {
             name: 'with-short-parameter',
             params: '-c'
@@ -24,6 +15,12 @@ module.exports = {
                 coverage: true
             }
         }
-    ]
+    ],
+
+    onStart: function (childProcess) {
+        childProcess.on('done', function () {
+            childProcess._testSnapshot('stdout', childProcess._stdout);
+        });
+    }
 
 }