+++ tests/sinon-sandboxing/config.js
@@ -1,0 +1,9 @@
+module.exports = {
+
+    onStart: function (childProcess) {
+        childProcess.on('done', function () {
+            childProcess._testSnapshot('stdout', childProcess._stdout);
+        });
+    }
+
+}