+++ tests/sinon-sandboxing/test.js
@@ -1,0 +1,9 @@
+class Test {
+
+    fn1() {
+        return true;
+    }
+
+}
+
+export default new Test();