+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 
 
+## [Unreleased]
+
+### Added
+- sinon sandboxing support
+
+
 ## [0.1.0] - 2016-10-05
 
 ### Added